api/requirements.txt
django~=4.2
djangorestframework
django-cors-headers~=4.6
environs